requirements.txt
# NFL RAG Application Dependencies

# Data Loading & Storage
nflreadpy>=0.2.0
pandas>=2.0.0
requests>=2.28.0
duckdb>=0.10.0
pyarrow>=14.0.0

# Text Processing
beautifulsoup4>=4.12.0

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0

# Scheduling (for automated updates)
schedule>=1.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # for testing FastAPI